requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
django-cors-headers==3.7.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.3
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.0.1
idna==2.10
mypy-extensions==0.4.3
numpy==1.20.1
openpyxl==3.0.7
pandas==1.2.3
pathspec==0.8.1
phonenumbers==8.12.20
Pillow==8.1.2
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
regex==2021.3.17
requests==2.25.1
serpy==0.3.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
XlsxWriter==1.3.7
